+++ module-stock-service/src/main/java/com/shop/stockservice/StockServiceApplication.java
@@ -2,11 +2,10 @@ package com.shop.stockservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @EnableFeignClients
-@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
+@SpringBootApplication
 public class StockServiceApplication {
 
     public static void main(String[] args) {
