+++ module-stock-service/src/main/java/com/shop/stockservice/stock/controller/InternalStockController.java
@@ -1,6 +1,6 @@
-package com.shop.stockservice.controller;
+package com.shop.stockservice.stock.controller;
 
-import com.shop.stockservice.service.StockService;
+import com.shop.stockservice.stock.service.StockService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
