+++ module-stock-service/src/main/java/com/shop/stockservice/stock/dto/response/StockResponse.java
@@ -1,4 +1,4 @@
-package com.shop.stockservice.dto.response;
+package com.shop.stockservice.stock.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
